Bump minimum supported node engine version, bump module to 0.3
Too many downstream dependencies don't support v0.8 and it's a dead
version at this point, so it makes sense to just raise the supported
engine value.

Bumping the supported engines is a big enough change to warrant a new
minor version, vs a patch version. It's still below 1.x, so semver
won't automatically scoop this up and break someone's build.

LICENSE is also updated in package.json and the year is brought up to date.